move erc20 (USDT) contract folder to under smoketest
