+++ contrib/localnet/orchestrator/smoketest/contracts/erc20/USDT.go
