+++ contrib/localnet/orchestrator/smoketest/contracts/erc20/bindings.go
