+++ contrib/localnet/orchestrator/smoketest/main.go
@@ -5,11 +5,11 @@ import (
 	"fmt"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/zeta-chain/zetacore/contracts/evm/erc20"
 	"github.com/zeta-chain/zetacore/contracts/evm/erc20custody"
 	"github.com/zeta-chain/zetacore/contracts/evm/zetaconnectoreth"
 	"github.com/zeta-chain/zetacore/contracts/evm/zetaeth"
 	contracts "github.com/zeta-chain/zetacore/contracts/zevm"
+	"github.com/zeta-chain/zetacore/contrib/localnet/orchestrator/smoketest/contracts/erc20"
 	fungibletypes "github.com/zeta-chain/zetacore/x/fungible/types"
 	"math/big"
 	"os"
