+++ contrib/localnet/orchestrator/smoketest/test_setup.go
@@ -6,11 +6,11 @@ import (
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/zeta-chain/zetacore/contracts/evm/erc20"
 	"github.com/zeta-chain/zetacore/contracts/evm/erc20custody"
 	"github.com/zeta-chain/zetacore/contracts/evm/zetaconnectoreth"
 	"github.com/zeta-chain/zetacore/contracts/evm/zetaeth"
 	contracts "github.com/zeta-chain/zetacore/contracts/zevm"
+	"github.com/zeta-chain/zetacore/contrib/localnet/orchestrator/smoketest/contracts/erc20"
 	fungibletypes "github.com/zeta-chain/zetacore/x/fungible/types"
 	"math/big"
 	"time"
